@@ -2,7 +2,7 @@ package info
 
 import (
 	"errors"
-	"io/ioutil"
+	"os"
 	"path/filepath"
 
 	"golang.org/x/mod/modfile"
@@ -13,7 +13,7 @@ var ErrModuleNameNotFound = errors.New("module name not found")
 // BuildName extracts the last part of the module by splitting on `/`
 // this last part is useful for name of the binary and other things.
 func BuildName() (string, error) {
-	content, err := ioutil.ReadFile("go.mod")
+	content, err := os.ReadFile("go.mod")
 	if err != nil {
 		return "", err
 	}
@@ -32,7 +32,7 @@ func BuildName() (string, error) {
 // from go.mod, it returns empty if there is
 // an issue reading the go.mod
 func ModuleName() string {
-	content, err := ioutil.ReadFile("go.mod")
+	content, err := os.ReadFile("go.mod")
 	if err != nil {
 		return ""
 	}
