@@ -2,7 +2,6 @@ package npm
 
 import (
 	"context"
-	"io/ioutil"
 	"os"
 	"testing"
 )
@@ -23,7 +22,7 @@ func TestBuildCmd(t *testing.T) {
 			commandNil:  false,
 			failMessage: "command should not be nil",
 			beforeFn: func() {
-				err := ioutil.WriteFile("package-lock.json", []byte{}, 0600)
+				err := os.WriteFile("package-lock.json", []byte{}, 0600)
 				if err != nil {
 					t.Error(err)
 				}
